Upgrades storybook to version 7.0.11. Fixes issues with storybook running. Adds build storybook to ui build github action.
