+++ src/ui/common/.babelrc
@@ -1,6 +1,7 @@
 {
   "presets": [
-    "@babel/preset-react"
+    "@babel/preset-react",
+    "@babel/preset-typescript"
   ],
   "plugins": [
     "@emotion/babel-plugin"
