+++ src/ui/common/.storybook/main.js
@@ -1,29 +1,21 @@
 module.exports = {
   module: {
-    rules: [
-      {
-        test: /\.(stories|story)\.mdx$/,
-        use: [
-          {
-            loader: require.resolve('@storybook/mdx2-csf/loader'),
-            options: {},
-          }
-        ]
-      }
-    ]
+    rules: [{
+      test: /\.(stories|story)\.mdx$/,
+      use: [{
+        loader: require.resolve('@storybook/mdx2-csf/loader'),
+        options: {}
+      }]
+    }]
   },
   stories: ["../src/**/*.stories.mdx", "../src/**/*.stories.@(js|jsx|ts|tsx)"],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-  ],
-  framework: "@storybook/react",
+  addons: ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions", "@storybook/addon-mdx-gfm"],
+  framework: {
+    name: "@storybook/react-webpack5",
+    options: {}
+  },
   features: {
     previewMdx2: true
-  },
-  core: {
-    builder: "@storybook/builder-webpack5",
   },
   typescript: {
     check: false,
@@ -39,10 +31,10 @@ module.exports = {
       // makes string and boolean types that can be undefined appear as inputs and switches
       shouldRemoveUndefinedFromOptional: true,
       // Filter out third-party props from node_modules except @mui packages
-      propFilter: (prop) =>
-        prop.parent
-          ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName)
-          : true,
-    },
+      propFilter: prop => prop.parent ? !/node_modules\/(?!@mui)/.test(prop.parent.fileName) : true
+    }
   },
+  docs: {
+    autodocs: true
+  }
 };
