+++ src/ui/common/package.json
@@ -24,8 +24,8 @@
     "analyze": "size-limit --why",
     "lint": "eslint '*/**/*.{js,ts,tsx}' --format table",
     "lint:fix": "eslint '*/**/*.{js,ts,tsx}' --format table --fix",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "resolutions": {
     "@types/react": "^18.0.8"
@@ -39,13 +39,14 @@
     "@fortawesome/free-brands-svg-icons": "^6.2.1",
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
+    "@hookform/devtools": "^4.3.1",
     "@hookform/resolvers": "^2.8.4",
-    "@tag0/use-text-width": "^1.2.0",
     "@monaco-editor/react": "^4.4.5",
     "@mui/lab": "5.0.0-alpha.117",
     "@mui/material": "5.11.6",
     "@mui/x-data-grid": "5.17.8",
     "@reduxjs/toolkit": "^1.9.3",
+    "@tag0/use-text-width": "^1.2.0",
     "@ungap/structured-clone": "^1.0.2",
     "@zip.js/zip.js": "^2.6.62",
     "autosuggest-highlight": "^3.3.4",
@@ -67,7 +68,6 @@
     "react-virtualized": "^9.22.3",
     "reactflow": "11.6.0",
     "unist-util-visit-parents": "^5.1.3",
-    "@hookform/devtools": "^4.3.1",
     "yup": "^0.32.11"
   },
   "husky": {
@@ -95,17 +95,18 @@
   "devDependencies": {
     "@babel/core": "^7.20.12",
     "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "7.21.5",
     "@mdx-js/react": "^2.2.1",
     "@size-limit/preset-small-lib": "^7.0.8",
-    "@storybook/addon-actions": "6.5.12",
-    "@storybook/addon-essentials": "6.5.12",
-    "@storybook/addon-interactions": "6.5.12",
-    "@storybook/addon-links": "6.5.12",
-    "@storybook/builder-webpack5": "6.5.12",
-    "@storybook/manager-webpack5": "6.5.12",
-    "@storybook/mdx2-csf": "^0.0.3",
-    "@storybook/react": "6.5.12",
-    "@storybook/testing-library": "0.0.13",
+    "@storybook/addon-actions": "7.0.11",
+    "@storybook/addon-essentials": "7.0.11",
+    "@storybook/addon-interactions": "7.0.11",
+    "@storybook/addon-links": "7.0.11",
+    "@storybook/addon-mdx-gfm": "^7.0.11",
+    "@storybook/mdx2-csf": "^1.1.0",
+    "@storybook/react": "7.0.11",
+    "@storybook/react-webpack5": "^7.0.11",
+    "@storybook/testing-library": "0.1.0",
     "@types/react": "^18.0.8",
     "@types/react-dom": "^18.0.5",
     "@typescript-eslint/eslint-plugin": "^5.49.0",
@@ -121,6 +122,7 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-storybook": "^0.6.12",
     "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^8.0.1",
     "parcel": "^2.8.3",
@@ -130,6 +132,7 @@
     "react-hook-form": "^7.43.9",
     "rollup-plugin-postcss": "^4.0.2",
     "size-limit": "^7.0.8",
+    "storybook": "^7.0.11",
     "tslib": "^2.4.0",
     "typescript": "^4.9.4",
     "vite": "^4.0.4",
